% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{compute.eBIC}
\alias{compute.eBIC}
\title{Extended Bayesian Information Criterion (eBIC) for pdRCON models}
\usage{
compute.eBIC(S, admm.out, n, gamma.eBIC = 0.5, max_iter = 5000, mle = TRUE)
}
\arguments{
\item{S}{a sample covariance matrix with the block structure described in \code{\link{pdglasso-package}}.}

\item{admm.out}{an object of class \code{ADMMoutput}, such as the output of a call to \code{\link{admm.pdglasso}}; see also \code{\link{pdRCON.select}}.}

\item{n}{the sample size.}

\item{gamma.eBIC}{a parameter governing the magnitude of the penalization term inside the criterion;
it ranges from 0 to 1, where 0 makes the eBIC equivalent to BIC, with 0.5 being the value
suggested by Feygel and Drton (2010).}

\item{max_iter}{an integer; maximum number of iterations for convergence.}

\item{mle}{a logical; if \code{TRUE}, the MLE are used otherwise
the pdglasso estimator is used; see the "Details" section below.}
}
\value{
A vector containing three elements:
\itemize{
\item the value of the eBIC,
\item the log-likelihood,
\item the number of parameters.
}
}
\description{
This function computes the value of the eBIC for a pdRCON model
from the output of \code{\link{admm.pdglasso}}; see Eq.1 of Feygel and Drton (2010).
}
\details{
\strong{Details on the specification of the argument \code{mle}}

The extended Bayesian Information Criterion implemented in this function requires that the parameters are estimated by maximum likelihood.
However, the computation of MLEs may considerably slow down the procedure and, more seriously, in the high-dimensional setting
MLEs may even not exist. For this reason, we provide the option that the eBIC is, naively, computed by using the pdglasso estimate.
}
\examples{
S <- cov(toy_data$sample.data)
admm.out <- admm.pdglasso(S, lambda1=4, lambda2=0.7)
compute.eBIC(S, admm.out, n=60, gamma.eBIC=0.5)

compute.eBIC(S, admm.out, n=60, gamma.eBIC=0.5, mle=FALSE)
}
\references{
Foygel, R. and Drton, M. (2010). Extended Bayesian information criteria for Gaussian graphical models. \emph{Advances in neural information processing systems}, 23.
}
