% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fMRI_parietal}
\alias{fMRI_parietal}
\title{fMRI dataset}
\format{
\subsection{\code{fMRI_parietal}}{

A \eqn{404 \times 10} dataframe with 404 observations referring to residuals from \eqn{p=5\times 2=10} time-series
after filtering through a score-driven model.
The residuals are associated to 10 different regions of interest of the parietal area of the brain, five on the left empisphere e the paired
homologous regions on the right hemisphere.
}
}
\usage{
fMRI_parietal
}
\description{
This dataset contains residuals obtained through score-driven models on the time-series of brain activity at rest for an individual,
measured on 10 regions of interest (ROI) of the brain in the parietal area (five on the left hemisphere and five paired regions on the right hemisphere).
Please check Sections 2 and 7 of Ranciati and Roverato (2021) for more information.
}
\keyword{datasets}
