% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{G.merge}
\alias{G.merge}
\title{Conversion from the multiple matrix representation of the model to the single matrix representation.}
\usage{
G.merge(X)
}
\arguments{
\item{X}{list with three upper triangular matrices with entries 0 and 1: G, G.sym and G.across, any of G.sym and G.across may be NULL}
}
\value{
a pXp symmetric matrix with entries 0, 1, and 2
}
\description{
This is the inverse of the function \link{G.split}, i.e. X is equal to \code{G.split(G.merge(X))}.
}
\examples{

# random generation of a list(G=G, G.sym=G.sym, G.across=G.across)

q <- 5 # this can be any integer
p <- q*2

g.p <- matrix(sample(c(0,1), size=p^2, replace=TRUE), nrow=p, ncol=p)
g.p[lower.tri(g.p, diag=TRUE)] <- 0

g.q1 <- matrix(sample(c(0,1), size=q^2, replace=TRUE), nrow=q, ncol=q)
g.q1[lower.tri(g.q1, diag = FALSE)] <- 0

g.q2 <- matrix(sample(c(0,1), size=q^2, replace=TRUE), nrow=q, ncol=q)
g.q2[lower.tri(g.q2, diag = TRUE)] <- 0
g.q2sym <- g.q2+t(g.q2)

g.p[1:q, 1:q] <- g.p[1:q, 1:q] *(1-g.q1)
g.p[(q+1):p, (q+1):p] <- g.p[(q+1):p, (q+1):p] *(1-g.q1)
g.p[1:q, (q+1):p] <- g.p[1:q, (q+1):p] * (1-g.q2sym)

# list obtained

X <- list(G=g.p, G.sym=g.q1, G.across=g.q2)

g  <- G.merge(X)
gs <- G.split(g)

identical(X, gs)

X <- list(G=g.p, G.sym=NULL, G.across=NULL)

g  <- G.merge(X)
gs <- G.split(g)

identical(X, gs)

}
