% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{G.split}
\alias{G.split}
\title{Conversion from the single matrix representation of the model to the multiple matrix representation.}
\usage{
G.split(g)
}
\arguments{
\item{g}{is a pXp symmetric matrix with entries 0, 1, and 2}
}
\value{
a list with three upper triangular matrices: G, G.sym and G.across with entries 0 and 1, any of G.sym and G.across may be NULL
}
\description{
This is the inverse of the function \code{\link[=G.merge]{G.merge()}}, i.e. g is equal to \code{G.merge(G.split(g))}.
}
\examples{

# see example in function [G.merge()].
}
