% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{get.pdColG}
\alias{get.pdColG}
\title{Build a graph from the output of a call to \code{\link{admm.pdglasso}}.}
\usage{
get.pdColG(admm.out, th1 = NULL, th2 = NULL, verbose = FALSE)
}
\arguments{
\item{admm.out}{An object of list type, that is the output of a call to the \code{\link{admm-pdglasso}} function.}

\item{th1}{(optional) A scalar, the threshold to identify edges in the graph; it must be non-negative.}

\item{th2}{(optional) A scalar, the threshold to identify coloured edges in the graph; it must be non-negative.}

\item{verbose}{(optional) if \code{TRUE} provides summary statistics of the graph.}
}
\value{
a list, containing:
\itemize{
\item g, the graph in matrix form.
\item dof, the degrees of freedom corresponding to the graph build under the pdglasso model provided.
}
}
\description{
Description here.
}
\examples{

S <- cov(toy_data)
mod.out <- admm.pdglasso(S)
get.pdColG(mod.out)
}
