% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{GGM.simulate}
\alias{GGM.simulate}
\title{Random simulation of Gaussian graphical models (GGMs)}
\usage{
GGM.simulate(
  p,
  concent.mat = TRUE,
  sample = TRUE,
  Sigma = NULL,
  sample.size = NULL,
  dens = 0.1
)
}
\arguments{
\item{p}{an even integer; the number of variables of the generated model; 
see the "Details" section below.}

\item{concent.mat}{a logical; if `TRUE` a
concentration matrix `K` adapted to `G` is generated.}

\item{sample}{a logical; if `TRUE` a sample from a
normal distribution with zero mean vector and concentration matrix `K`
is generated.}

\item{Sigma}{a pxp positive definite matrix; the matrix
argument of the function [`rWishart`], which is used as starting point in
the random generation of the concentration matrix, as described in the
"Details" section of the function [`pdRCON.simulate`]. If `NULL`
the identity matrix is used.}

\item{sample.size}{a positive integer; the size of the randomly generated sample. If `NULL`
then the sample size is set to 3xp.}

\item{dens}{a value between zero and one used to specify the sparsity
degree of the generated graph, as described in the "Details" section of the
function [`pdRCON.simulate`].}
}
\value{
A list with the following components:

 * `G` a randomly generated (symmetric) adjacency matrix of an undirected
 graph on `p` vertices.

* `K`  a randomly generated concentration matrix adapted to `G`  
    if `concent.mat=TRUE`, or `NULL` otherwise.

* `sample.data` a data frame randomly generated from a multivariate normal distribution
  with mean vector zero and concentration matrix `K`  if both 
  `sample=TRUE` and `concent.mat=TRUE`, or `NULL` otherwise.


Note that the variable are named `V1,...,Vp`.
}
\description{
Randomly generates a GGM and, more specifically, the undirected graph `G` 
representing the model, a concentration matrix `K` adapted to `G`, 
that is a positive definite matrix 
whose zero pattern is specified by the missing edges of `G`, and a random sample
from a multivariate normal distribution with zero mean vector and covariance
matrix `Sigma`, that is the inverse of `K`.
}
\details{
A GGM is a pdRCON model with no parametric symmetries, and the
  purpose of this function is that of providing a simplified call to the
  function [`pdRCON.simulate`], with the appropriate choice of arguments
  required to simulate a GGM. Note, however, that pdRCON models make
  sense only if the number of variables p is even, and this requirement is
  (unnecessarily) retained here.
}
\examples{

# generates distribution and data from a GGM on 20 variables and graph density equal to 0.2
p <- 20
n <-100
set.seed(1234)
GenMod <- GGM.simulate(p, sample.size=n, dens=0.20)

# check graph sparsity degree
n.edges <- sum(GenMod$G[upper.tri(GenMod$G)])
n.edges/(p*(p-1)/2)

# check positive definiteness
min(eigen(GenMod$K)$values)

# computation of the partial correlation matrix
R <- -cov2cor(GenMod$K)
diag(R) <- 1
}
