% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-functions.R
\name{is.pdRCON.mle}
\alias{is.pdRCON.mle}
\title{Check if a concentration matrix satisfies the likelihood equations.}
\usage{
is.pdRCON.mle(K.mle, pdColG, S, toll = 1e-08, print.checks = TRUE)
}
\arguments{
\item{K.mle}{candidate mle of \eqn{K} to be checked.}

\item{pdColG}{matrix representing a colored graph for paired data;
see [`pdglasso-package`] for details.}

\item{S}{a sample covariance matrix with the block structure describe
in [`pdglasso-package`].}

\item{toll}{threshold to check whether a value is equal to zero.}

\item{print.checks}{a logical (default `TRUE`).}
}
\value{
a logical equal to `TRUE` if `K.mle` satisfies the likelihood equations
and `FALSE` otherwise. If `print.checks = TRUE` the values used for the
checking are printed.
}
\description{
Checks if a matrix is the maximum likelihood estimate of the concentration
matrix of a pdRCON model represented by the colored graph for paired data
`pdColG`, computed from the covariance matrix `S`.
}
\details{
In order to check if `K.mle` is the maximum likelihood estimate, computed
from `S`, under the model represented by `pdColG` the following quantities
are considered:

* the values of `K.mle` that are expected to be zero;

* the differences between the entries of `K.mle` which are expected to be equal;

* the differences between the relevant sums of the entries of `solve(K.mle)`
  and `S` which are expected to be equal, so as to satisfy the
  likelihood equations.

* all the absolute quantities of the previous three points are divided
  by the \eqn{\ell_2} norm of the non-zero entries of `K.mle` so as to
  obtain relative quantities.

  Then if all the above quantities are smaller than `toll` the check is
  considered successful and a `TRUE` is returned.
}
\keyword{internal}
