% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalANDutils.R
\name{lams.max}
\alias{lams.max}
\title{Maximum theoretical values of lambda1 and lambda2}
\usage{
lams.max(S)
}
\arguments{
\item{S}{a sample covariance matrix.}
}
\value{
A vector of two elements, `lambda1.max` and `lambda2.max`.
}
\description{
Computes the maximum theoretical values `lambda1.max` and `lambda2.max` of `lambda1` and `lambda2`, respectively. 
For every `lambda1` greater than `lambda1.max` the pdglasso estimator is a diagonal matrix whereas 
for every `lambda2` greater than `lambda2.max` the pdglasso estimator is fully symmetric.
}
\examples{
S <- cov(toy_data$sample.data)
lams.max(S)
}
