% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{max.lams}
\alias{max.lams}
\title{Compute maximum theoretical values for \code{lambda1} and \code{lambda2}.}
\usage{
\method{max}{lams}(S)
}
\arguments{
\item{S}{a covariance matrix.}
}
\value{
a vector of two elements.
}
\description{
Computes the maximum values for \code{lambda1} and \code{lambda2} such that:
\itemize{
\item if max of \code{lambda1} is used, the estimated concentration matrix will be diagonal;
\item if max of \code{lambda2} is used, the estimated concentration matrix will be fully symmetric.
}
}
\examples{
S <- cov(toy_data$sample.data$)
max.lams(S)
}
