% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdglasso-package.R
\docType{package}
\name{pdglasso-package}
\alias{pdglasso}
\alias{pdglasso-package}
\title{pdglasso: Graphical Lasso for Coloured Gaussian Graphical Models for Paired Data}
\description{
RCON models are Colured Gaussian Graphical Models with equality "R"estrictions on "CON"centration values (Højsgaard and Lauritzen, 2008). The package pdglasso implements methods for RCON models for paired data (pdRCON models), including an Alternating Direction Method of Multipliers (ADMM) algorithm for solving the penalized likelihood problem introduced by Ranciati and Roverato (2024). It also implements functions for the computation of maximum likelihood estimates and for generating simulated pdRCON models and datasets.
}
\details{
An RCON model for paired data (pdRCON model) is a coloured Gaussian Graphical
Model (GGM) where the \eqn{p} variables are partitioned into a Left block
\eqn{L} and a Right block \eqn{R}. The two blocks are not independent, every
variable in the left block has an homologous variable in the right block and
certain types of equality "R"estrictions on the entries of the
"CON"centration matrix \eqn{K} are allowed. A pdRCON model is represented by
a Coloured Graph for Paired Data (pdColG) with a vertex for every variable
and where every vertex and edge is either *coloured* or *uncoloured*. More
details on the equality constraints of pdRCON models, on submodel classes
of interest and on the usage of the package are given in the following.
}
\section{pdRCON models - terminology and relevant submodel classes}{


A pdRCON model is a Gaussian Graphical Model (GGM) with additional equality
restrictions on the entries of the concentration matrix. In the paired data
framework, there are three different types of equality restrictions of
interest, identified by the names *vertex*, *inside-block edge* and
*across-block edge*, respectively. Relevant submodel classes can be specified both
by allowing different combinations of restriction types and by forcing
different types of fully symmetric structures. In this package, different
submodel classes are identified by the arguments `type` and `force.symm` and
models are represented by coloured graphs for paired data encoded in the form
of a pdColG matrix.

  * Every variable in \eqn{L} has an homologous variable in \eqn{R} and the
corresponding diagonal entries of \eqn{K} can be constrained to have equal
value. Such entries of  \eqn{K} are represented by *coloured vertices* of the
independence graph whereas the unconstrained diagonal entries are represented
by  *uncoloured vertices*. Different types of submodel classes of interest
may be obtained by (i) not allowing coloured vertices, (ii) allowing both
coloured and uncoloured vertices and (iii) allowing only coloured vertices.

  * For every pair of variables in \eqn{L} there exists an homologous pair of
variables in \eqn{R}, thereby identifying a pair of homologous edges. If both
edges are present in the graph the corresponding off-diagonal entries of
\eqn{K} can be constrained to have equal value. These type of edges are
referred to as *coloured inside-block edges*. Different types of
submodel classes of interest may be obtained by (i) not allowing coloured
inside-block edges, (ii) allowing both coloured and uncoloured inside-block
edges and (iii) allowing only coloured inside-block edges.

  * We say that two variables are
  *across-block* if one variable belongs to \eqn{L}
and the other to \eqn{R}. For every pair of non-homologous across-block
variables there exists an homologous pair across-block variables, thereby
identifying a pair of homologous edges. If both edges are present in the
graph the corresponding off-diagonal entries of \eqn{K} can be constrained to
have equal value. These type of edges are referred to as *coloured
across-block edges*. Different types of submodel classes of interest may be
obtained by (i) not allowing coloured across-block edges, (ii) allowing both
coloured and uncoloured across-block edges and (iii) allowing only coloured
across-block edges, with the exception of edges joining a variable in \eqn{L}
with its homologous in \eqn{R}.

A pair of homologous edges which are  both present
in the graph form what we call a *structural symmetry*. A structural symmetry can be 
either *uncoloured* or *coloured* and the latter is also called a *parametric symmetry*.
Accordingly, a pair of homologous coloured vertices is called a *vertex symmetry*. 
It is worth remarking that, strictly speaking, a pair of homologous missing edges 
may be regarded as a parametric symmetry because the associated concentrations have the
same, zero, value. However, we deem it convenient to use this terminology only to refer to  
present edges.
}

\section{Use of the arguments type and force.symm to specify submodel classes}{


 The functions of this package make it possible to specify different types of
 pdRCON submodel classes of interest through the arguments `type` and
 `force.symm` which can both take as value any subvector of the character
 vector `c("vertex", "inside.block.edge", "across.block.edge")`; note that
 the names of the components can be abbreviated down, up to the first letter
 only, and are not case-sensitive. The argument `type` cannot be `NULL` and:

 * If `type` contains the string `"vertex"` then vertex symmetries are
 allowed and, if in addition also `force.symm` contains the string
 `"vertex"`, then all vertices are coloured.

 * If `type` contains the string `"inside.block.edge"` then coloured inside-block symmetries are allowed and, if in addition also `force.symm`
 contains the string `"inside.block.edge"`, then only coloured edges are
 allowed inside blocks.

 * If `type` contains the string `"across.block.edge"` then coloured across-block symmetries are allowed and, if in addition also `force.symm`
 contains the string `"across.block.edge"`, then only coloured edges are
 allowed across blocks,  with the exception of edges joining a variable in
 \eqn{L} with its homologous in \eqn{R}.


Note that  `force.symm` is a, possibly  `NULL`, subvector of  `type`.
Elements of `force.symm` which are not elements of `type` are ignored.
}

\section{Variables position and block structure of sample covariance matrices}{


  The functions of this package assume that the positions occupied by
  variables follow certain rules. More specifically, the positions from
  \eqn{1} to \eqn{q=p/2} correspond to the first group of variables, say
  \eqn{L}, whereas the positions from \eqn{q+1} to \eqn{p} are associated
  with the second group of variables, \eqn{R}. Furthermore, the variables of
  the first group are ordered in the same way as those of the second group,
  in the sense that for every \eqn{i=1,\ldots, q} the variable in position
  \eqn{i} is homologous to the variable in position \eqn{q+i}. Hence, for
  instance, the functions that receive in input a sample covariance matrix
  assume that the rows and columns of this matrix are ordered according to
  these rules, so that it can be partitioned into four \eqn{q\times q}
  submatrices naturally associated with the inside- and across-block
  components. The same is true for the `pdColG` matrix described below.
}

\section{Model representation through a pdColG matrix}{


 Every pdRCON model is uniquely represented by a Coloured Graph for Paired
 Data (pdColG) implemented in the form of an object of class `pdColG`
 that is a \eqn{p\times p} symmetric matrix
 where every entry is one of the values 0, 1 or 2, as follows:

 * The diagonal entries of a pdColG matrix are all equal to either 1,
 for uncoloured vertices, or 2, for coloured vertices.

 * The off-diagonal entries of a pdColG matrix are equal to 0 for missing
 edges and either 1 or 2 for present edges, where the value 2 is used to
 encode coloured edges.
}

\references{
Højsgaard, S. and Lauritzen, S. L. (2008). Graphical Gaussian models with edge and vertex symmetries. *Journal of the Royal Statistical Society Series B: Statistical Methodology*, *70*(5), 1005-1027.

Ranciati, S. and Roverato, A., (2024). On the application of Gaussian graphical models to paired data problems. *Statistics and Computing*, *34*(6), 1-19.

Ranciati, S., Roverato, A. and Luati, A. (2021). Fused graphical lasso for brain networks with symmetries. *Journal of the Royal Statistical Society Series C: Applied Statistics*, *70*(5), 1299-1322.

Roverato, A. and Nguyen, D.N. (2024). Exploration of the search space of Gaussian graphical models for paired data. *Journal of Machine Learning Research*, *25*(92), 1-41.
}
\author{
Saverio Ranciati (ORCID: 0000-0001-7880-9465), Alberto Roverato (ORCID: 0000-0001-7984-3593),
Anna Vesely (ORCID: 0000-0001-6696-2390), and Peter Majcen (ORCID: 0009-0009-7020-7246)
\cr **Maintainer**: Anna Vesely <anna.vesely2@unibo.it>
}
