% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{pdRCON.simulate}
\alias{pdRCON.simulate}
\title{Random simulation of pdRCON models}
\usage{
pdRCON.simulate(
  p,
  concent.mat = TRUE,
  sample = TRUE,
  Sigma = NULL,
  sample.size = NULL,
  type = c("vertex", "inside.block.edge", "across.block.edge"),
  force.symm = NULL,
  dens = 0.1,
  dens.vertex = NULL,
  dens.inside = NULL,
  dens.across = NULL,
  verbose = TRUE
)
}
\arguments{
\item{p}{an even integer; the number of variables of the generated model.}

\item{concent.mat}{a logical; if \code{TRUE} a
concentration matrix \code{K} with the zero structure and symmetries encoded by \code{pdColG} is generated.}

\item{sample}{a logical; if \code{TRUE} a sample from a
normal distribution with zero mean vector and concentration matrix \code{K}
is generated.}

\item{Sigma}{a pxp positive definite matrix; the matrix
argument of the \code{\link{rWishart}} function, which is used as starting point in
the random generation of the concentration matrix, as described in the
"Details" section below. If \code{NULL}
the identity matrix is used.}

\item{sample.size}{a positive integer; the size of the randomly generated sample. If \code{NULL}
then the sample size is set to 3xp.}

\item{type, force.symm}{two subvectors of \code{c("vertex", "inside.block.edge", "across.block.edge")} which
identify the pdRCON submodel class of interest; see \code{\link{pdglasso-package}} for details.}

\item{dens, dens.vertex, dens.inside, dens.across}{four values between zero and
one used to specify the sparsity degree of the generated graph, as
explained in the "Details" section below. The default \code{dens.vertex=NULL} is
equivalent to \code{dens.vertex=dens}, and similarly for \code{dens.inside} and
\code{dens.across}.}

\item{verbose}{a logical; if \code{TRUE} the pdRCON submodel class considered, as
specified by the arguments \code{type} and \code{force.symm}, is returned as printed
output in the console.}
}
\value{
A list with the following components:
\itemize{
\item \code{pdColG} a randomly generated a matrix of class \code{pdColG} representing a pdRCON model; see \code{\link{pdglasso-package}} for details.
\item \code{K} a randomly generated concentration matrix with the zero structure and symmetries encoded by \code{pdColG}
if \code{concent.mat=TRUE}, or \code{NULL} otherwise.
\item \code{sample.data}  a data frame randomly generated from a multivariate normal distribution
with mean vector zero and concentration matrix \code{K}   if both
\code{sample=TRUE} and \code{concent.mat=TRUE}, or \code{NULL} otherwise.
}

Note that the variable in \eqn{L} are named \verb{L1,...,Lq} and variables in
\eqn{R} are are named \verb{R1,...,Rq} where  \code{Li} is homologous to  \code{Ri} for
every i=1,...,q.
}
\description{
Randomly generates a pdRCN model and, more specifically, the pdColG matrix representing the model,
a concentration matrix \code{K} with the same zero pattern and equality constraints encoded by \code{pdColG},  and a
random sample from a multivariate normal distribution with zero mean vector
and covariance matrix \code{Sigma}, that is the inverse of \code{K}.
}
\details{
\strong{Details on the sparsity degree of the generated graph}

The argument \code{dens.vertex} specifies the proportion of coloured vertices
among the \eqn{p} vertices. This is used if the string "vertex" is a
component of \code{type} but not of \code{force.symm}. The string "vertex" not being
a component of \code{type} is equivalent to \code{dens.vertex=0} whereas the string
"vertex" being a component of both \code{type} and \code{force.symm} is equivalent to
\code{dens.vertex=1}.

The argument \code{dens.inside} specifies the proportion of coloured symmetric
inside-block edges among the \eqn{q(q-1)} inside-block
edges, where \eqn{q=p/2}, . This is used if the string "inside.block.edge" is a component of
\code{type}, otherwise it is equivalent to \code{dens.inside=0}. The overall density
of inside-block edges is obtained by the sum of the densities of coloured
and uncoloured inside-block edges. This is  a value between \code{dens.inside}
and \code{dens.inside+dens}. Furthermore, it is exactly equal to \code{dens} if
"inside.block.edge" is not a component of \code{type} and to \code{dens.inside} if
"inside.block.edge" is a component of both \code{type} and \code{force.symm}.

The argument \code{dens.across} specifies the proportion of coloured symmetric
across-block edges among the potentially coloured \eqn{q(q-1)} across-block
edges. This is used if the string "across.block.edge" is a component
of \code{type}, otherwise it is equivalent to \code{dens.across=0}. The overall
density of across-block edges is obtained by the sum of the densities of
coloured and uncoloured across-block edges. This is  a value between
\code{dens.across} and \code{dens.across+dens}. Furthermore, it is exactly equal  to
\code{dens} if "across.block.edge" is not a component of \code{type}.

The argument \code{dens} specifies the density of uncoloured edges. Note that
the algorithm generates uncoloured edges first, which may be overwritten by
coloured edges. For this reason the actual density of uncoloured edges is
typically smaller than \code{dens}.

\strong{Details on the generating process of the concentration matrix}

The concentration matrix is obtained by first generating a random Wishart
matrix with matrix parameter \code{Sigma} and \eqn{p} degrees of freedom, which
represents an initial unconstrained covariance matrix. This is inverted and
adapted to a suitable coloured graph for paired data with sparsity degree
according to the  \code{dens.xxx} arguments.
}
\examples{

# generates a pdRCON model on 10 variables in the form of a pdColG matrix

set.seed(123)
rpdColG <- pdRCON.simulate(10, concent=FALSE, sample=FALSE, dens=0.25)$pdColG
rpdColG

# generates a distribution from a pdRCON model on 20 variables, a concentration matrix
# for this model and a sample of size 50
# all vertices are coloured and no coloured across-block edge is allowed

set.seed(123)
GenMod <- pdRCON.simulate(20, type=c("v", "i"), force.symm=c("v"), sample.size=50, dens=0.20)

summary(GenMod$pdColG)

# computation of the partial correlation matrix
R <- -cov2cor(GenMod$K)
diag(R) <- 1
}
