% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy dataset generated through [`pdRCON.simulate`] function}
\format{
## `toy_data`
A list with the following components:

* `pdColG` a matrix representing a coloured graph for paired data; see [`pdglasso-package`] for details.
* `K`, a concentration matrix adapted to `pdColG`.
* `sample.data`, a data frame with 250 rows and 20 columns from a multivariate normal distribution
with zero mean vector and concentration matrix `K`.
}
\usage{
toy_data
}
\description{
Data simulated by the function [`pdRCON.simulate`] with parameters:
* `p=20`
* `type=c("v", "i","a")`
* `force.symm=NULL`
* `dens=0.5`
* `Sigma` a \eqn{p \times p} matrix with unitary diagonal and 0.5 on off-diagonal elements.
}
\keyword{datasets}
