% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm-adaptive.R
\name{admm.inner}
\alias{admm.inner}
\title{Title}
\usage{
admm.inner(
  X,
  U,
  rho1,
  lambda1,
  lambda2,
  rho2 = 1,
  varying.rho2 = TRUE,
  max_iter_int = 1000,
  eps.abs = 1e-12,
  eps.rel = 1e-08,
  verbose_int = FALSE,
  n.row.F,
  acr.type
)
}
\arguments{
\item{rho1}{A scalar; tuning parameter of the ADMM algorithm to be used for
the outer loop. It must be strictly positive.}

\item{lambda1}{A non-negative scalar (or vector) penalty that encourages
sparsity in the concentration matrix. If a vector is provided, it should
match the appropriate length, i.e.}

\item{lambda2}{A non-negative scalar (or vector) penalty that encourages
equality constraints in the concentration matrix. If a vector is provided,
it should match the appropriate length, i.e.}

\item{rho2}{A scalar; tuning parameter of the ADMM algorithm to be used for
the inner loop. It must be strictly positive.}

\item{varying.rho2}{A boolean value; if \code{TRUE} the parameter rho2 is updated
iteratively to speed-up convergence.}

\item{eps.abs}{A scalar; the absolute precision required for the computation
of primal and dual residuals of the ADMM algorithm.}

\item{eps.rel}{A scalar; the relative precision required for the computation
of primal and dual residuals of the ADMM algorithm.}
}
\description{
Title
}
