% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm-lasso.R
\name{admm.pdglasso}
\alias{admm.pdglasso}
\title{ADMM graphical lasso algorithm for pdRCON models}
\usage{
admm.pdglasso(
  S,
  lambda1,
  lambda2,
  type = c("vertex", "inside.block.edge", "across.block.edge"),
  force.symm = NULL,
  X.init = NULL,
  rho1 = 1,
  rho2 = 1,
  varying.rho1 = TRUE,
  varying.rho2 = TRUE,
  max_iter = 5000,
  eps.abs = 1e-06,
  eps.rel = 1e-06,
  rcpp = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{S}{a sample covariance matrix with the block structure
described in \code{\link{pdglasso-package}}.}

\item{lambda1, lambda2}{two non-negative scalar penalties that encourage
sparsity (\code{lambda1}) and parametric symmetries (\code{lambda2}) in the concentration matrix.}

\item{type, force.symm}{two subvectors of \code{c("vertex", "inside.block.edge", "across.block.edge")} which identify the pdRCON submodel class of interest; see
\code{\link{pdglasso-package}} for details.}

\item{X.init}{a  matrix of the same dimension as \code{S} to be used as starting point of the ADMM. If \code{NULL} a default value is used.}

\item{rho1, rho2}{two positive scalars; tuning parameters of the outer and inner loop, respectively, of the ADMM.}

\item{varying.rho1, varying.rho2}{two logicals; if \code{TRUE} the parameters \code{rho1} and \code{rho2}, respectively, are updated
iteratively to speed-up convergence.}

\item{max_iter}{an integer; maximum number of iterations for convergence.}

\item{eps.abs, eps.rel}{two  positive scalars; the absolute and relative tolerance, respectively, for the computation
of primal and dual feasibility tolerances of the ADMM.}

\item{rcpp}{a logical; if \code{TRUE}, computations are performed using the Rcpp (C++) implementation;
if \code{FALSE}, a pure (slower) R implementation is used.}

\item{verbose}{a logical; if \code{TRUE} the pdRCON submodel class considered, as
specified by the arguments \code{type} and \code{force.symm}, is returned as printed
output on the console.}
}
\value{
A object of class \code{ADMMoutput} that is a list with the following components:
\itemize{
\item \code{X} the estimated concentration matrix.
\item \code{acronyms} a vector of strings identifying the pdRCON submodel class
considered as identified  by the arguments \code{type} and \code{force.symm}.
\item \code{internal.par} a list of internal parameters passed to the function at the
call, as well as convergence information.
}
}
\description{
Estimate of a concentration matrix within the pdRCON
submodel class identified by the arguments \code{type} and \code{force.symm},
based on the pdglasso method with penalty terms \code{lambda1} and \code{lambda2}
(Ranciati and Roverato, 2024). Optimization is implemented by an
ADMM algorithm (see Boyd \emph{et. al.} 2011). The output is an object of
class \code{ADMMoutput} and the user may then call the function \code{\link{pdColG.get}} to obtain the Coloured Graph
for Paired Data (pdColG) representing the selected model.
}
\examples{

# computation of the sample covariance
S <- cov(toy_data$sample.data)

# model with all types of symmetries allowed and no full symmetry required
admm.out <- admm.pdglasso(S, lambda1=4, lambda2=0.7)
G <- pdColG.get(admm.out)
summary(G)

# model with no across-block symmetries allowed and full vertex-symmetry required
admm.out <- admm.pdglasso(S, lambda1=4, lambda2=0.7, type=c("v", "i"),
            force.symm = c("v"), verbose=TRUE)
G <- pdColG.get(admm.out)
summary(G)

}
\references{
Ranciati, S. and Roverato, A., (2024). On the application of Gaussian graphical models to paired data problems. \emph{Statistics and Computing}, \emph{34}(6), 1-19.

Boyd, S., Parikh, N., Chu, E., Borja, P. and Eckstein, J. (2011). Distributed optimization and statistical learning via the alternating direction method of multipliers. \emph{Foundations and Trends in Machine learning}, \emph{3}(1), 1-122.
}
