% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm-adaptive.R
\name{admm.pdglasso}
\alias{admm.pdglasso}
\title{Estimates a concentration matrix under the pdglasso model using adaptive ADMM algorithm.}
\usage{
admm.pdglasso(
  S,
  lambda1 = 1,
  lambda2 = 1e-04,
  type = c("vertex", "inside.block.edge", "across.block.edge"),
  force.symm = NULL,
  X.init = NULL,
  rho1 = 1,
  rho2 = 1,
  varying.rho1 = TRUE,
  varying.rho2 = TRUE,
  max_iter = 1000,
  eps.abs = 1e-12,
  eps.rel = 1e-12,
  verbose = FALSE
)
}
\arguments{
\item{S}{A \eqn{p \times p} covariance (or correlation) matrix.}

\item{lambda1}{A non-negative scalar (or vector) penalty that encourages
sparsity in the concentration matrix. If a vector is provided, it should match the appropriate length, i.e.}

\item{lambda2}{A non-negative scalar (or vector) penalty that encourages
equality constraints in the concentration matrix. If a vector is provided, it should match the appropriate length, i.e.}

\item{type}{The type of equality constraints to be imposed; zero, one or more available options
can be selected among:
\itemize{
\item "vertex", symmetries are imposed on the diagonal entries of the concentration matrix.
\item "inside.block.edge", symmetries are imposed between elements of the LL and RR block the concentration matrix.
\item "across.block.edge", symmetries are imposed between elements of the LR and RL block the concentration matrix.
}}

\item{verbose}{}
}
\description{
Description here.
}
\examples{
!!! Create fake dataset
S <- cov(toy.data)
admm.pdglasso(S)
}
