% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-functions.R
\name{compute.eBIC}
\alias{compute.eBIC}
\title{Compute the extended Bayesian Information Criterion (eBIC).}
\usage{
compute.eBIC(S, mod, n, gamma.eBIC = 0.5, max_iter = 1000)
}
\arguments{
\item{S}{a sample covariance (or correlation) matrix with the block structure described in \code{\link{pdglasso-package}}.}

\item{mod}{a list, the output object of a call to \code{\link{admm.pdglasso}}}

\item{n}{the sample size of the data used to compute the sample covariance matrix S.}

\item{gamma.eBIC}{a parameter governing the magnitude of the penalization term inside the criterion; it ranges from 0 to 1, where 0 makes the eBIC equivalent to BIC, and 0.5 being the suggested default value.}

\item{max_iter}{an integer; maximum number of iterations to be run in case
the algorithm does not converge; passed to \code{\link{pdRCON.mle}}.}
}
\value{
A vector containing three elements:
\itemize{
\item the value of the eBIC,
\item the log-likelihood,
\item and the estimated number of degrees of freedom.
}
}
\description{
The function computes the value of the eBIC for a given model and gamma value, for the purpose
of model selection (see Eq.1 of Feygel & Drton, 2010).
}
\examples{
S <- cov(toy_data$sample.data)
mod <- admm.pdglasso(S, lambda1=1, lambda2=0.5)
compute.eBIC(S,mod,n=60,gamma.eBIC=0.5)
}
\references{
Foygel, R., Drton, M. (2010). Extended Bayesian information criteria for Gaussian graphical models. \emph{Advances in neural information processing systems}, 23. \url{https://proceedings.neurips.cc/paper/2010/file/072b030ba126b2f4b2374f342be9ed44-Paper.pdf}
}
