% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-functions.R
\name{compute.eBIC}
\alias{compute.eBIC}
\title{Compute the extended Bayesian Information Criterion (eBIC).}
\usage{
compute.eBIC(S, mod, n, gamma.eBIC = 0.5)
}
\arguments{
\item{S}{a \eqn{p \times p} covariance (or correlation) matrix.}

\item{mod}{a list, the output object of a call to \code{\link{admm.pdglasso}}}

\item{n}{the sample size of the data used to compute the sample covariance matrix S.}

\item{gamma.eBIC}{a parameter needed to compute the eBIC; ranges from 0 to 1, where 0 makes the eBIC equivalent to BIC.}
}
\value{
A vector containing three elements:
\itemize{
\item the value of the eBIC,
\item the log-likelihood,
\item and the estimated number of degrees of freedom.
}
}
\description{
The function computes the value of the eBIC for a given model and gamma value, for the purpose
of model selection.
}
\examples{
S <- cov(toy_data$sample.data)
mod <- admm.pdglasso(S, lambda1=1, lambda2=0.5)
compute.eBIC(S,mod,n=60,gamma.eBIC=0.5)
}
