% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fMRI_parietal}
\alias{fMRI_parietal}
\title{fMRI Dataset}
\format{
\subsection{\code{fMRI_parietal}}{

A \eqn{404 \times 10} matrix.
}
}
\usage{
fMRI_parietal
}
\description{
Put reference here.
}
\keyword{datasets}
