% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{GGM.simulate}
\alias{GGM.simulate}
\title{Random simulation of Gaussian graphical models (GGMs)}
\usage{
GGM.simulate(
  p,
  concent.mat = TRUE,
  sample = TRUE,
  Sigma = NULL,
  sample.size = NULL,
  dens = 0.1
)
}
\arguments{
\item{p}{an even integer, the number of variables of the generated model;
see the "Details" section below.}

\item{concent.mat}{a logical, if \code{TRUE} a
concentration matrix \code{K} adapted to \code{G} is generated.}

\item{sample}{a logical, if \code{TRUE} a sample from a
normal distribution with zero mean vector and concentration matrix \code{K}
is generated.}

\item{Sigma}{a \eqn{p\times p} positive definite matrix. This is the matrix
argument of the \code{\link{rWishart}} function, which is used as starting point in
the random generation of the concentration matrix, as described in the
"Details" section of the function \code{\link{pdRCON.simulate}}. If \code{NULL}
the identity matrix is used.}

\item{sample.size}{a positive integer, size of the randomly generated sample. If \code{NULL}
then \code{sample.size=3*p}.}

\item{dens}{a value between zero and one used to specify the sparsity
degree of the generated graph, as described in the "Details" section of the
function \code{\link{pdRCON.simulate}}.}
}
\value{
A list with the following components:
\itemize{
\item \code{G} a randomly generated (symmetric) adjacency matrix of an undirected
graph on \code{p} vertices.
\item \code{K} either a randomly generated concentration matrix adapted to \code{G} or
\code{NULL} if \code{concent.mat=FALSE}.
\item \code{sample.data}  either a data frame randomly generated from a multivariate normal distribution
with mean vector zero and concentration matrix \code{K} or \code{NULL} if either
\code{sample=FALSE} or \code{concent.mat=FALSE}.
}

Note that the variable are named \verb{V1,...,Vp}.
}
\description{
Randomly generates a GGM and, more specifically, the undirected graph \code{G}
representing the model, a concentration matrix \code{K} adapted to \code{G},
that is a positive definite matrix
whose zero pattern is specified by the missing edges of \code{G}, and a random sample
from a multivariate normal distribution with zero mean vector and covariance
matrix \code{Sigma}, that is the inverse of \code{K}.
}
\details{
A GGM is a pdRCON model with no parametric symmetries, and the
purpose of this function is that of providing a simplified call to the
function \code{\link{pdRCON.simulate}}, with the appropriate choice of arguments
required to simulate a GGM. Note, however, that pdRCON models make
sense only if the number of variables \code{p} is even, and this requirement is
(unnecessarily) retained here.
}
\examples{

# generates distribution and data from a GGM on 20 variables and graph density equal to 0.2
p <- 20
n <-100
set.seed(1234)
GenMod <- GGM.simulate(20, sample.size=n, dens=0.20)

# check graph sparsity degree
n.edges <- sum(GenMod$G[upper.tri(GenMod$G)])
n.edges/(p*(p-1)/2)

# check positive definiteness
min(eigen(GenMod$K)$values)

# computation of the partial correlation matrix
R <- -cov2cor(GenMod$K)
diag(R) <- 1
}
