% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul-functions.R
\name{GGM.simulate}
\alias{GGM.simulate}
\title{Random simulation of Gaussian graphical models (GGMs)}
\usage{
GGM.simulate(
  p,
  concent.mat = TRUE,
  sample = TRUE,
  Sigma = NULL,
  sample.size = NULL,
  dens = 0.1
)
}
\arguments{
\item{p}{an even integer, that is the number of vertices of the generated
undirected graph \code{G}.}

\item{concent.mat}{a logical (default \code{TRUE}) indicating whether a
concentration matrix \code{K} adapted to \code{G} should be generated.}

\item{sample}{a logical (default \code{TRUE}) indicating whether a sample from a
normal distribution with zero mean vector and concentration matrix \code{K}
should be generated.}

\item{Sigma}{a \eqn{p\times p} positive definite matrix. This is the matrix
argument of the \code{\link{rWishart}} function, which is used as starting point in
the random generation of the concentration matrix, as described in the
details section of the function \code{\link{pdRCON.simulate}}. The default \code{NULL} is
equivalent to the identity matrix \code{Sigma=diag(p)}.}

\item{sample.size}{size of the randomly generated sample. The default \code{NULL}
is equivalent to \code{sample.size=3*p}.}

\item{dens}{a value between zero and one used to specify the sparsity
degree of the generated graph, as described in the details section of the
function \code{\link{pdRCON.simulate}}.}
}
\value{
A list with the following components:
\itemize{
\item \code{G} a randomly generated matrix encoding an undirected
graph on \eqn{p} vertices. More specifically, \code{G} is a \code{pdColG} graph with
no entry equal to 2, and it differs from the adjacency matrix of the
undirected graph because its diagonal entries are equal to one.
\item \code{K} a randomly generated concentration matrix adapted to \code{G}.
\item \code{sample.data} a randomly generated sample form a multivariate normal distribution
with mean vector zero and concentration matrix \code{K}.
}

Note that the variable are named \verb{V1,...,Vp}.
}
\description{
Randomly generates an undirected graph \eqn{\mathcal{G}}, a
concentration matrix \eqn{K} adapted to \eqn{\mathcal{G}} and a random sample
from a multivariate normal distribution with zero mean vector and covariance
matrix \eqn{\Sigma=K^{-1}}.
}
\details{
A GGM is a pdRCON model with no parametric symmetries, and the
purpose of this function is that of providing a simplified call to the
function \code{\link{pdRCON.simulate}}, with the appropriate choice of arguments
required to simulate from GGMs. Note, however, that pdRCON models make
sense only if the number of variables \code{p} is even, and this requirement is
(unnecessarily) retained here.
}
\examples{

# generates distribution and data from a GGM on 20 variables and graph density equal to 0.2
p <- 20
n <-100
set.seed(1234)
GenMod <- GGM.simulate(20, sample.size=n, dens=0.20)

# check graph sparsity degree
n.edges <- sum(GenMod$G[upper.tri(GenMod$G)])
n.edges/(p*(p-1)/2)

# check positive definiteness
min(eigen(GenMod$K)$values)

# computation of the partial correlation matrix
R <- -cov2cor(GenMod$K)
diag(R) <- 1

}
