% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalANDutils.R
\name{lams.max}
\alias{lams.max}
\title{Maximum theoretical values of \code{lambda1} and \code{lambda2}}
\usage{
lams.max(S)
}
\arguments{
\item{S}{a sample covariance matrix.}
}
\value{
A vector of two elements, \code{lambda1.max} and \code{lambda2.max}.
}
\description{
Computes the maximum theoretical values \code{lambda1.max} and \code{lambda2.max} of \code{lambda1} and \code{lambda2}, respectively.
For every \code{lambda1} greater than \code{lambda1.max} the pdglasso estimator is a diagonal matrix whereas
for every \code{lambda2} greater than \code{lambda2.max} the pdglasso estimator is fully symmetric.
}
\examples{
S <- cov(toy_data$sample.data)
lams.max(S)
}
