% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{lams.max}
\alias{lams.max}
\title{Compute maximum theoretical values for `lambda1` and `lambda2`.}
\usage{
lams.max(S)
}
\arguments{
\item{S}{a covariance matrix.}
}
\value{
A vector of two elements.
}
\description{
Computes the maximum values for `lambda1` and `lambda2` such that:
* if max of `lambda1` is used, the estimated concentration matrix will be diagonal;
* if max of `lambda2` is used, the estimated concentration matrix will be fully symmetric.
}
\examples{
S <- cov(toy_data$sample.data)
lams.max(S)
}
