% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{pdColG.get}
\alias{pdColG.get}
\title{pdColG matrix from the output of a call to \code{\link{admm.pdglasso}}}
\usage{
pdColG.get(admm.out, th1 = NULL, th2 = NULL, model.dimension = FALSE)
}
\arguments{
\item{admm.out}{an object of class \code{ADMMoutput}.}

\item{th1, th2}{two positive scalars, the thresholds to identify edges
(\code{th1}) and coloured edges (\code{th2}) in the graph, if \code{NULL}
a default value is used; as describe in the "Details" section below. .}

\item{model.dimension}{a logical, if \code{TRUE} the number of parameters of the model is returned.}
}
\value{
Either an object of class \code{pdColG}, if \code{model.dimension=FALSE}, or a list with the following components if \code{model.dimension=TRUE}:
\itemize{
\item \code{pdColG} an object of class \code{pdColG}, that is a matrix representing a coloured graph for paired data.
\item \code{n.par} the number of parameters of the pdRCON model.
}
}
\description{
This function returns the pdColG matrix representing the pdRCON model
resulting from a call to \code{\link{admm.pdglasso}}; see also \code{\link{pdRCON.select}}.
}
\details{
\strong{Details of the specification of the argument \code{th1} and \code{th2}}

Due to finite-precision arithmetic in computing, the fitted concentration matrix obtained from the ADMM has no exact zeros and no exact identical
pairs of concentration values. Hence, \code{th1} and \code{th2} provide tollerances for values and differences, respectively, to be regarded as close enough to
zero. If  \code{NULL} tollerances are automatically derived from tollerance values used to check convergence of the ADMM.
The function \code{\link{plot.ADMMoutput}} allows to visualize the role played by the default threshold as well as to specify
facilitate the specification of suitable values for the personalized thresholds \code{th1} and \code{th2}.
}
\examples{

S <- cov(toy_data$sample.data)
admm.out <- admm.pdglasso(S, lambda1=4, lambda2=0.7)
pdColG.get(mod.out)
}
