% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{pdColG.summarize}
\alias{pdColG.summarize}
\title{Summary statistics for coloured graphs for paired data}
\usage{
pdColG.summarize(pdColG, print.summary = TRUE)
}
\arguments{
\item{pdColG}{a coloured graph for paired data.}

\item{print.summary}{logical  (default \code{TRUE}) indicating whether a summary should be printed.}
}
\value{
a list
}
\description{
This function
}
\examples{
#
pdColG.summarize(toy_data$pdColG)
}
