% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-functions.R
\name{pdRCON.fit}
\alias{pdRCON.fit}
\title{Fit and select a coloured GGM for paired data according to eBIC
criterion.}
\usage{
pdRCON.fit(
  S,
  n,
  n.l1 = 15,
  n.l2 = 15,
  gamma.eBIC = 0.5,
  type = c("vertex", "inside.block.edge", "across.block.edge"),
  force.symm = NULL,
  X.init = NULL,
  rho1 = 1,
  rho2 = 1,
  varying.rho1 = TRUE,
  varying.rho2 = TRUE,
  max_iter = 1000,
  eps.abs = 1e-12,
  eps.rel = 1e-12,
  verbose = FALSE,
  print.type = TRUE
)
}
\arguments{
\item{S}{a sample covariance (or correlation) matrix with the block structure
described in \code{\link{pdglasso-package}}.}

\item{n}{the sample size of the data used to compute the sample covariance matrix S.}

\item{n.l1}{the number of values in the grid of candidates for \code{lambda1}.}

\item{n.l2}{the number of values in the grid of candidates for \code{lambda2}.}

\item{gamma.eBIC}{the parameter for the eBIC computation. gamma=0 is equivalent to BIC.}

\item{type, force.symm}{two subvectors of \code{c("vertex", "inside.block.edge", "across.block.edge")} which identify the pdRCON submodel class of interest; see
\code{\link{pdglasso-package}} for details.}

\item{X.init}{(optional) a \eqn{p \times p} initial guess for the
concentration matrix and/or starting solution for the ADMM algorithm.}

\item{rho1}{a scalar; tuning parameter of the ADMM algorithm to be used for
the outer loop. It must be strictly positive.}

\item{rho2}{a scalar; tuning parameter of the ADMM algorithm to be used for
the inner loop. It must be strictly positive.}

\item{varying.rho1}{a logical; if \code{TRUE} the parameter rho1 is updated
iteratively to speed-up convergence.}

\item{varying.rho2}{a logical; if \code{TRUE} the parameter rho2 is updated
iteratively to speed-up convergence.}

\item{max_iter}{an integer; maximum number of iterations to be run in case
the algorithm does not converge.}

\item{eps.abs}{a scalar; the absolute precision required for the computation
of primal and dual residuals of the ADMM algorithm.}

\item{eps.rel}{a scalar; the relative precision required for the computation
of primal and dual residuals of the ADMM algorithm.}

\item{verbose}{a logical; if \code{TRUE} the progress (and internal convergence of
inner loop) is shown in the console while the algorithm is running.}

\item{print.type}{a logical; if \code{TRUE} the pdRCON submodel class considered, as
specified by the arguments \code{type} and \code{force.symm} - is returned as printed
output in the console.}
}
\value{
A list with the following components:
\itemize{
\item \code{model} the final model.
\item \code{pdColG} the associated coloured graph.
\item \code{best.lambdas} the selected values of \code{lambda1} and \code{lambda2} according to eBIC criterion.
\item \code{l1.path} a matrix containing the grid values for \code{lambda1} as well as quantities used in eBIC computation.
\item \code{l2.path} a matrix containing the grid values for \code{lambda2} as well as quantities used in eBIC computation.
}
}
\description{
Performs a sequence of calls to \code{\link{admm.pdglasso}} providing two grids of
values for \code{lambda1} and \code{lambda2}. First, a grid search conditional on
\code{lambda2}=0 is run to select the best \code{lambda1} value among the candidates
(according to eBIC); conditional on the best \code{lambda1}, a similar search is
performed for \code{lambda2}. The output is the select model, given by the
estimated concenration matrix and corresponding graph.
}
\examples{
S <- cov(toy_data$sample.data)
pdRCON.fit(S,n=60)
}
