% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-functions.R
\name{pdRCON.mle}
\alias{pdRCON.mle}
\title{Maximum likelihood estimation}
\usage{
pdRCON.mle(S, pdColG, eps.rel = 1e-06, eps.abs = 1e-06, max_iter = 5000)
}
\arguments{
\item{S}{a sample covariance matrix with the block structure described in \code{\link{pdglasso-package}}.}

\item{pdColG}{pdColG a matrix representing a coloured graph for paired data; see \code{\link{pdglasso-package}} for details.}

\item{eps.rel}{a scalar; the relative precision required for the computation
of primal and dual residuals of the ADMM algorithm.}

\item{eps.abs}{a scalar; the absolute precision required for the computation
of primal and dual residuals of the ADMM algorithm.}

\item{max_iter}{an integer; maximum number of iterations to be run in case
the algorithm does not converge.}
}
\value{
Either a matrix, that is the maximum likelihood estimate of
\eqn{K=\Sigma^{-1}} under the pdRCON model represented by \code{pdColG}, or
\code{NULL} if the maximum likelihood estimate does not exist.
}
\description{
Computes the maximum likelihood estimate of the concentration matrix of a pdRCON model.
}
\details{
If the sample covariance matrix is not full-rank, then it is
possible that the maximum likelihood estimate does not exist. The maximum
likelihood estimate is computed by running the function \code{\link{admm.pdglasso}}
with suitable penalties and, if it does not exist, then the ADMM algorithm
fails to converge, a warning is produced and a \code{NULL} is returned.
}
\examples{

S <- var(toy_data$sample.data)
K.hat <- pdRCON.mle(S, toy_data$pdColG)

}
