% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-functions.R
\name{pdRCON.mle}
\alias{pdRCON.mle}
\title{Maximum likelihood estimate}
\usage{
pdRCON.mle(S, pdColG, verbose = TRUE)
}
\arguments{
\item{S}{a sample variance and covariance matrix.}

\item{pdColG}{a coloured graph for paired data.}
}
\value{
the m.l.e. of the concentration matrix \eqn{\Sigma^{-1}}.
}
\description{
Computes the m.l.e. of the concentration matrix of a colured graphical model
for paired data.
}
\examples{
#

}
