% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{pdRCON.select}
\alias{pdRCON.select}
\title{Selection and estimate of a pdRCON model according to eBIC}
\usage{
pdRCON.select(
  S,
  n,
  lams = NULL,
  gamma.eBIC = 0.5,
  type = c("vertex", "inside.block.edge", "across.block.edge"),
  force.symm = NULL,
  X.init = NULL,
  rho1 = 1,
  rho2 = 1,
  varying.rho1 = TRUE,
  varying.rho2 = TRUE,
  max_iter = 5000,
  eps.abs = 1e-08,
  eps.rel = 1e-08,
  verbose = TRUE,
  print.type = TRUE,
  mle.estimate = TRUE
)
}
\arguments{
\item{S}{a sample covariance matrix with the block structure
described in [`pdglasso-package`].}

\item{n}{the sample size.}

\item{lams}{a 2x4 numeric matrix that specifies the path of lambda values, if `NULL` default values are
used, as explained in the "Details" section below.}

\item{gamma.eBIC}{a value between zero and one; the magnitude of the penalization term of the eBIC; see  [`compute.eBIC`] for details.}

\item{type, force.symm}{two subvectors of `c("vertex", "inside.block.edge",
"across.block.edge")` which identify the pdRCON submodel class of interest; see
[`pdglasso-package`] for details.}

\item{X.init}{a  matrix of the same dimension as `S` to be used as starting point of the ADMM. If `NULL` a default value is used.}

\item{rho1, rho2}{two positive scalars; tuning parameters of the outer and inner loop, respectively, of the ADMM.}

\item{varying.rho1, varying.rho2}{two logicals; if `TRUE` the parameters `rho1` and `rho2`, respectively, are updated
iteratively to speed-up convergence.}

\item{max_iter}{an integer; maximum number of iterations for convergence.}

\item{eps.abs, eps.rel}{two  positive scalars; the absolute and relative tollerance, respectively, for the computation
of primal and dual feasibility tollerances of the ADMM.}

\item{verbose}{a logical; if `TRUE` provides a visual update in the console about the grid search over `lambda1` and `lambda2`}

\item{print.type}{a logical; if `TRUE` the pdRCON submodel class considered, as
specified by the arguments `type` and `force.symm`, is returned as printed
output in the console.}

\item{mle.estimate}{a logical; if `TRUE`, compute eBIC via the MLE,
if `FALSE` the pdglasso estimator is used; see [`compute.eBIC`] for details.}
}
\value{
A list with the following components:

* `model` selected model; object of class `ADMMoutput` output of [`admm.pdglasso`].

* `lambda.grid` the grid of values used for `lambda1` and `lambda2.`

* `best.lambdas` the selected values of `lambda1` and `lambda2` according to eBIC.

* `l1.path` a matrix containing the path values for `lambda1` as well as relevant quantities of the eBIC.

* `l2.path` a matrix containing the path values for `lambda2` as well as relevant quantities of the eBIC.

* `time.exec` total execution time for the called function.

A warning is produced if at least one run of the algorithm for the grid
searches has resulted in non-convergence (status can be checked by inspecting
`l1.path` and `l2.path`); see the "Details" section of the function  [`compute.eBIC`].
}
\description{
Performs a sequence of calls to [`admm.pdglasso`] on a sequence of
values for `lambda1` and `lambda2`. More specifically, firstly a path of `lambda1` values, with 
`lambda2`=0, is considered to select a "best" `lambda1` value according to eBIC. 
Next, a path of `lambda2` values is considered, with `lambda1` fixed to its previously selected value, 
and the final model is selected according to eBIC.
The user can then call [`pdColG.get`] to obtain the Coloured Graph
for Paired Data (pdColG) representing the selected model.
}
\details{
**Details on the specification of the argument lams**

The argument `lams` is a 2x4 matrix used to specified the grid of penalty terms.
The first row of `lams` refers to `lambda1` and second row to
`lambda2`. For each row of the `lams` matrix, entries are: the minimum and maximum value of the path (columns 1 and 2); 
 the number of points of the path (column 3); `0` for linear spacing and  
`1` logarithmic spacing (column 4). If `lams=NULL` the default values are computed as follows: maximum lambda values are computed from [`lams.max`]
and paths of 20 points are used form max.lams/20 to max.lams, with logarithm spacing.
}
\examples{
S <- cov(toy_data$sample.data)
sel.mod <- pdRCON.select(S, n=60)
sel.mod$l1.path
sel.mod$l2.path
plot(sel.mod$model)
pdColG.get(sel.mod$model)
}
