% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{pdRCON.select}
\alias{pdRCON.select}
\title{Selection and estimate of a pdRCON model according to eBIC}
\usage{
pdRCON.select(
  S,
  n,
  lams = NULL,
  gamma.eBIC = 0.5,
  type = c("vertex", "inside.block.edge", "across.block.edge"),
  force.symm = NULL,
  X.init = NULL,
  rho1 = 1,
  rho2 = 1,
  varying.rho1 = TRUE,
  varying.rho2 = TRUE,
  max_iter = 5000,
  eps.abs = 1e-08,
  eps.rel = 1e-08,
  verbose = TRUE,
  print.type = TRUE,
  mle.estimate = TRUE
)
}
\arguments{
\item{S}{a sample covariance matrix with the block structure
described in \code{\link{pdglasso-package}}.}

\item{n}{the sample size.}

\item{lams}{a 2x4 numeric matrix to specify the path of lambda values, if \code{NULL} default values are
used; see the "Details" section below for additional information.}

\item{gamma.eBIC}{parameter of eBIC with \code{gamma.eBIC=0} corresponding to the classical BIC; see \code{\link{compute.eBIC}} for details..}

\item{type, force.symm}{two subvectors of \code{c("vertex", "inside.block.edge", "across.block.edge")} which identify the pdRCON submodel class of interest; see
\code{\link{pdglasso-package}} for details.}

\item{X.init}{a \eqn{p \times p}
concentration matrix to be used as starting point of the ADMM. If \code{NULL} a default value is used.}

\item{rho1}{a postive scalar; tuning parameter of the ADMM algorithm to be used for
the outer loop.}

\item{rho2}{a positive scalar; tuning parameter of the ADMM algorithm to be used for
the inner loop.}

\item{varying.rho1}{a logical; if \code{TRUE} the parameter \code{rho1} is updated
iteratively to speed-up convergence.}

\item{varying.rho2}{a logical; if \code{TRUE} the parameter \code{rho2} is updated
iteratively to speed-up convergence.}

\item{max_iter}{an integer; maximum number of iterations for convergence.}

\item{eps.abs}{a  positive scalar; the absolute tollerance for the computation
of primal and dual feasibility tollerances of the ADMM.}

\item{eps.rel}{a positive scalar; the relative tollerance for the computation
of primal and dual feasibility tollerances of the ADMM.}

\item{verbose}{a logical; if \code{TRUE} provides a visual update in the console about the grid search over \code{lambda1} and \code{lambda2}}

\item{print.type}{a logical; if \code{TRUE} the pdRCON submodel class considered, as
specified by the arguments \code{type} and \code{force.symm}, is returned as printed
output in the console.}

\item{mle.estimate}{a logical; if \code{TRUE}, compute eBIC via the MLE,
if \code{FALSE} the pdglasso estimator is used; see \code{\link{compute.eBIC}} for details.}
}
\value{
A list with the following components:
\itemize{
\item \code{model} selected model; object of class \code{ADMMoutput} output of \code{\link{admm.pdglasso}}.
\item \code{lambda.grid} the grid of values used for \code{lambda1} and \code{lambda2.}
\item \code{best.lambdas} the selected values of \code{lambda1} and \code{lambda2} according to eBIC.
\item \code{l1.path} a matrix containing the path values for \code{lambda1} as well as relevant quantities of the eBIC.
\item \code{l2.path} a matrix containing the path values for \code{lambda2} as well as relevant quantities of the eBIC.
\item \code{time.exec} total execution time for the called function.
}

A warning is produced if at least one run of the algorithm for the grid
searches has resulted in non-convergence (status can be checked by inspecting
\code{l1.path} and \code{l2.path}).
}
\description{
Performs a sequence of calls to \code{\link{admm.pdglasso}} on a sequence of
values for \code{lambda1} and \code{lambda2}. More specifically, firstly a path of \code{lambda1} values, with
\code{lambda2}=0, is considered to select a "best" \code{lambda1} value according to eBIC.
Next, a path of \code{lambda2} values is considered, with \code{lambda1} fixed to its previously selected value,
and the final model is selected according to eBIC.
The user may then call \code{\link{pdColG.get}} to obtain the Coloured Graph
for Paired Data (pdColG) representing the selected model.
}
\details{
\strong{Details on the specification of the argument \code{lams}}

The argument \code{lams} is a 2x4 matrix used to specified the grid of penalty terms.
The first row of \code{lams} refers to \code{lambda1} and second row to
\code{lambda2}. For each row of the \code{lams} matrix entries are: the minimum and maximum value of the path (columns 1 and 2);
the number of points of the path (column 3); \code{0} for linear spacing and
\code{1} logarithmic spacing (column 4). If \code{lams=NULL} the default values are computed as follows: maximum lambda values are computed from \code{\link{lams.max}}
and paths of 20 points are used form max.lams/20 to max.lams, with logarithm spacing.
}
\examples{
S <- cov(toy_data$sample.data)
sel.mod <- pdRCON.select(S, n=60)
sel.mod$l1.path
sel.mod$l2.path
plot(sel.mod$model)
pdColG.get(sel.mod$model)
}
