% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{pdRCON.simulate}
\alias{pdRCON.simulate}
\title{Random simulation of pdRCON models}
\usage{
pdRCON.simulate(
  p,
  concent.mat = TRUE,
  sample = TRUE,
  Sigma = NULL,
  sample.size = NULL,
  type = c("vertex", "inside.block.edge", "across.block.edge"),
  force.symm = NULL,
  dens = 0.1,
  dens.vertex = NULL,
  dens.inside = NULL,
  dens.across = NULL,
  print.type = TRUE
)
}
\arguments{
\item{p}{an even integer; the number of variables of the generated model.}

\item{concent.mat}{a logical; if `TRUE` a
concentration matrix `K` with the zero structure and symmetries encoded by `pdColG` is generated.}

\item{sample}{a logical; if `TRUE` a sample from a
normal distribution with zero mean vector and concentration matrix `K`
is generated.}

\item{Sigma}{a pxp positive definite matrix; the matrix
argument of the [`rWishart`] function, which is used as starting point in
the random generation of the concentration matrix, as described in the
"Details" section below. If `NULL`
the identity matrix is used.}

\item{sample.size}{a positive integer; the size of the randomly generated sample. If `NULL`
then the sample size is set to 3xp.}

\item{type, force.symm}{two subvectors of `c("vertex", "inside.block.edge", "across.block.edge")` which
identify the pdRCON submodel class of interest; see [`pdglasso-package`] for details.}

\item{dens, dens.vertex, dens.inside, dens.across}{four values between zero and
one used to specify the sparsity degree of the generated graph, as
explained in the "Details" section below. The default `dens.vertex=NULL` is
equivalent to `dens.vertex=dens`, and similarly for `dens.inside` and
`dens.across`.}

\item{print.type}{a logical; if `TRUE` the pdRCON submodel class considered, as
specified by the arguments `type` and `force.symm`, is returned as printed
output in the console.}
}
\value{
A list with the following components:

 * `pdColG` a randomly generated a matrix of class `pdColG` representing a pdRCON model; see [`pdglasso-package`] for details.

*  `K` a randomly generated concentration matrix with the zero structure and symmetries encoded by `pdColG`    
    if `concent.mat=TRUE`, or `NULL` otherwise. 

* `sample.data`  a data frame randomly generated from a multivariate normal distribution
  with mean vector zero and concentration matrix `K`   if both 
  `sample=TRUE` and `concent.mat=TRUE`, or `NULL` otherwise.

Note that the variable in \eqn{L} are named `L1,...,Lq` and variables in
\eqn{R} are are named `R1,...,Rq` where  `Li` is homologous to  `Ri` for
every i=1,...,q.
}
\description{
Randomly generates a pdRCN model and, more specifically, the pdColG matrix representing the model, 
a concentration matrix `K` with the same zero pattern and equality constraints encoded by `pdColG`,  and a
random sample from a multivariate normal distribution with zero mean vector
and covariance matrix `Sigma`, that is the inverse of `K`.
}
\details{
**Details on the sparsity degree of the generated graph**

  The argument `dens.vertex` specifies the proportion of coloured vertices
  among the \eqn{p} vertices. This is used if the string "vertex" is a
  component of `type` but not of `force.symm`. The string "vertex" not being
  a component of `type` is equivalent to `dens.vertex=0` whereas the string
  "vertex" being a component of both `type` and `force.symm` is equivalent to
  `dens.vertex=1`.

  The argument `dens.inside` specifies the proportion of coloured symmetric
  inside-block edges among the \eqn{q(q-1)} inside-block
  edges, where \eqn{q=p/2}, . This is used if the string "inside.block.edge" is a component of
  `type`, otherwise it is equivalent to `dens.inside=0`. The overall density
  of inside-block edges is obtained by the sum of the densities of coloured
  and uncoloured inside-block edges. This is  a value between `dens.inside`
  and `dens.inside+dens`. Furthermore, it is exactly equal to `dens` if
  "inside.block.edge" is not a component of `type` and to `dens.inside` if
  "inside.block.edge" is a component of both `type` and `force.symm`.

  The argument `dens.across` specifies the proportion of coloured symmetric
  across-block edges among the potentially coloured \eqn{q(q-1)} across-block
  edges. This is used if the string "across.block.edge" is a component
  of `type`, otherwise it is equivalent to `dens.across=0`. The overall
  density of across-block edges is obtained by the sum of the densities of
  coloured and uncoloured across-block edges. This is  a value between
  `dens.across` and `dens.across+dens`. Furthermore, it is exactly equal  to
  `dens` if "across.block.edge" is not a component of `type`.

  The argument `dens` specifies the density of uncoloured edges. Note that
  the algorithm generates uncoloured edges first, which may be overwritten by
  coloured edges. For this reason the actual density of uncoloured edges is
  typically smaller than `dens`.

  **Details on the generating process of the concentration matrix**

  The concentration matrix is obtained by first generating a random Wishart
  matrix with matrix parameter `Sigma` and \eqn{p} degrees of freedom, which
  represents an initial unconstrained covariance matrix. This is inverted and
  adapted to a suitable coloured graph for paired data with sparsity degree
  according to the  `dens.xxx` arguments.
}
\examples{

# generates a pdRCON model on 10 variables in the form of a pdColG matrix

set.seed(123)
rpdColG <- pdRCON.simulate(10, concent=FALSE, sample=FALSE, dens=0.25)$pdColG
rpdColG

# generates a distribution from a pdRCON model on 20 variables, a concentration matrix
# for this model and a sample of size 50
# all vertices are coloured and no coloured across-block edge is allowed

set.seed(123)
GenMod <- pdRCON.simulate(20, type=c("v", "i"), force.symm=c("v"), sample.size=50, dens=0.20)

summary(GenMod$pdColG)

# computation of the partial correlation matrix
R <- -cov2cor(GenMod$K)
diag(R) <- 1
}
