% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul-functions.R
\name{pdRCON.simulate}
\alias{pdRCON.simulate}
\title{Title}
\usage{
pdRCON.simulate(
  p,
  concent.mat = TRUE,
  sample = TRUE,
  Sigma = NULL,
  sample.size = NULL,
  type = c("vertex", "inside.block.edge", "across.block.edge"),
  force.symm = NULL,
  dens = 0.1,
  dens.vertex = NULL,
  dens.inside = NULL,
  dens.across = NULL
)
}
\arguments{
\item{p}{number of variables.}

\item{concent.mat}{a logical (default \code{TRUE}) indicating whether a
concentration matrix should be generated.}

\item{sample}{a logical (default \code{TRUE}) indicating whether a sample form a
normal distribution with zero mean vector and the generated concentration
matrix should be generated.}

\item{Sigma}{x}

\item{sample.size}{sample size with default value equal to \eqn{3p}.}

\item{type}{x}

\item{force.symm}{x}

\item{dens}{x}

\item{dens.vertex}{x}

\item{dens.inside}{x}

\item{dens.across}{x}
}
\value{
this function
}
\description{
Title
}
\examples{
#
}
