% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul-functions.R
\name{pdRCON.simulate}
\alias{pdRCON.simulate}
\title{Random simulation of pdRCON models}
\usage{
pdRCON.simulate(
  p,
  concent.mat = TRUE,
  sample = TRUE,
  Sigma = NULL,
  sample.size = NULL,
  type = c("vertex", "inside.block.edge", "across.block.edge"),
  force.symm = NULL,
  dens = 0.1,
  dens.vertex = NULL,
  dens.inside = NULL,
  dens.across = NULL
)
}
\arguments{
\item{p}{an even integer, that is the number of vertices of the generated
coloured graph for paired data \code{pdColG}.}

\item{concent.mat}{a logical (default \code{TRUE}) indicating whether a
concentration matrix \code{K} adapted to \code{pdColG} should be generated.}

\item{sample}{a logical (default \code{TRUE}) indicating whether a sample from a
normal distribution with zero mean vector and concentration matrix \code{K}
should be generated.}

\item{Sigma}{a \eqn{p\times p} positive definite matrix. This the matrix
argument of the \code{\link{rWishart}} function, which is used as starting point in
the random generation of the concentration matrix, as described in the
details section below. The default \code{NULL} is equivalent to the identity
matrix \code{Sigma=diag(p)}.}

\item{sample.size}{size of the randomly generated sample. The default \code{NULL}
is equivalent to \code{sample.size=3*p}.}

\item{type, force.symm}{two subvectors of \code{c("vertex", "inside.block.edge", "across.block.edge")} which
identify the pdRCON (sub)model of interest; see \code{\link{pdglasso}} for details.}

\item{dens, dens.vertex, dens.inside, dens.across}{four values between zero and
one used to specify the sparsity degree of the generated graph, as
described in the details section below. The default \code{dens.vertex=NULL} is
equivalent to \code{dens.vertex=dens}, and similarly for \code{dens.inside} and
\code{dens.across}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{pdColG} the randomly generated coloured graph for paired data on \eqn{p} vertices; see \code{\link{pdglasso}} for details.
\item \code{K} the randomly generated concentration matrix adapted to \code{pdColG}.
\item \code{sample.data} the randomly generated sample form a multivariate normal distribution
with mean vector zero and concentration matrix \code{K}.

Note that the variable in \eqn{L} are are named \verb{L1,...,Lq} and variables
in \eqn{R} are are named \verb{R1,...,Rq} where  \code{Li} is homologous to  \code{Ri} for
every i=1,...,q.
}
}
\description{
Randomly generates a coloured graph for paired data \eqn{\mathcal{G}}, a
concentration matrix \eqn{K} adapted to \eqn{\mathcal{G}} and a random sample
from a multivariate normal distribution with zero mean vector and covariance
matrix \eqn{\Sigma=K^{-1}}.
}
\details{
\strong{Details on the sparsity degree of the generated graph}

The argument \code{dens.vertex} specifies the proportion of coloured vertices
among the \eqn{p} vertices. This is used if the string "vertex" is a
component of \code{type} but not of \code{force.symm}. The string "vertex" not being
a component of \code{type} is equivalent to \code{dens.vertex=0} whereas the string
"vertex" being a component of both \code{type} and \code{force.symm} is equivalent to
\code{dens.vertex=1}.

The argument \code{dens.inside} specifies the proportion of coloured symmetric
inside block edges among the \eqn{q(q-1)/2}, with \eqn{q=p/2}, inside block
edges. This is used if the string "inside.block.edge" is a component of
\code{type}, otherwise it is equivalent to \code{dens.inside=0}. The overall density
of inside block edges is obtained by the sum of the densities of coloured
and uncoloured inside block edges. This is  a value between \code{dens.inside}
and \code{dens.inside+dens}. Furthermore, it is exactly equal to \code{dens} if
"inside.block.edge" is not a component of \code{type} and to \code{dens.inside} if
"inside.block.edge" is a component of both \code{type} and \code{force.symm}.

The argument \code{dens.across} specifies the proportion of coloured symmetric
across block edges among the potentially coloured \eqn{q(q-1)/2} across
block edges. This is used if the string "across.block.edge" is a component
of \code{type}, otherwise it is equivalent to \code{dens.across=0}. The overall
density of across block edges is obtained by the sum of the densities of
coloured and uncoloured across block edges. This is  a value between
\code{dens.across} and \code{dens.across+dens}. Furthermore, it is exactly equal  to
\code{dens} if "across.block.edge" is not a component of \code{type}.

The argument \code{dens} specifies the density of uncoloured edges. Note that
the algorithm generates uncoloured edges first, which may be overwritten by
coloured edges. For this reason the actual density of uncoloured edges is
typically smaller than \code{dens}.

\strong{Details on the generating process of the concentration matrix}

The concentration matrix is obtained by first generating a random Wishart
matrix with matrix parameter \code{Sigma} and \eqn{p} degrees of freedom, which
represents an initial unconstrained covariance matrix. This is inverted and
adapted to a suitable coloured graph for paired data with sparsity degree
according to the  \code{dens.xxx} arguments.
}
\examples{

# generates a pdRCON model on 10 variables in the form of a pdColG matrix

set.seed(123)
pdRCON.model <- pdRCON.simulate(10, concent=FALSE, sample=FALSE, dens=0.25)$pdColG

# generates a pdRCON model on 20 variables, a concentration matrix
# for this model and a sample of size 50
# all vertices are coloured and no coloured across block edge is allowed

set.seed(123)
GenMod <- pdRCON.simulate(20, type=c("v", "i"), force.symm=c("v"), sample.size=50, dens=0.20)

}
