% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalANDutils.R
\name{plot.pdColG}
\alias{plot.pdColG}
\title{Visual representation of a coloured graph for paired data}
\usage{
\method{plot}{pdColG}(
  x,
  y = "everything",
  which.sym = c("structural", "parametric"),
  asym.edges = TRUE,
  export.plot = FALSE,
  fancy = TRUE,
  ...
)
}
\arguments{
\item{x}{a matrix of class \code{pdColG}; see \code{\link{pdglasso-package}} for details.}

\item{y}{a string; allows the user to specify which portion of the graph (a block) is to be plotted: possible mutually exclusive options are "left", "right", "across", "everything";  default is "everything", which means the entire graph is plotted.}

\item{which.sym}{a string or a vector of strings; specifies which kind of symmetries are plotted (structural, parametric or both).}

\item{asym.edges}{a logical; if \code{TRUE}, Asymmetric edges are plotted otherwise their symbol is suppressed}

\item{export.plot}{a logical; if \code{TRUE}, a .pdf file is produced in the working director instead of plotting the graph as a new panel/window.}

\item{fancy}{a logical value; if \code{TRUE}, symbols in the plot are used from the
Latin1 encoding for characters; set to FALSE if characters are not properly
displayed, so symbols are reverted to latin letters.}

\item{...}{additional graphical parameters. Currently not used; reserved for future extensions.}
}
\value{
Either a plot within the running R session or a .pdf file saved in the working directory if the option export.plot is set to \code{TRUE}.
}
\description{
Heatmap-style graphical representation of an object of class \code{pdColG},
such as that raturned by a call to  \code{\link{pdColG.get}}.
}
\examples{
plot(toy_data$pdColG)
}
