% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalANDutils.R
\name{summary.pdColG}
\alias{summary.pdColG}
\title{Structural properties of a coloured graph for paired data}
\usage{
\method{summary}{pdColG}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{a matrix of class `pdColG`; see [`pdglasso-package`] for details.}

\item{verbose}{a logical; if `TRUE` the summary is printed on the console.}

\item{...}{not used. Included for compatibility with the generic `summary()` method.}
}
\value{
An invisible list with the following components:

* `overall` a list with the number of vertices, edges and free parameters of the associated pdRCON model.

* `vertex`  a list with the number of coloured vertices.

* `inside`  a list with the number of inside-block edges, the number of uncolored symmetric and coloured inside block edges.

* `across`  a list with the number of across-block edges, the number of uncolored symmetric and coloured across block edges.
}
\description{
Summary statistics relative to the structural properties 
of an object of class `pdColG`, such as that raturned by a call to  [`pdColG.get`].
}
\examples{
summary(toy_data$pdColG)

model.summary <- summary(toy_data$pdColG, verbose=FALSE)
model.summary

}
