% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy dataset generated through \code{\link{pdRCON.simulate}} function}
\format{
\subsection{\code{toy_data}}{

A list with the following components:
\itemize{
\item \code{pdColG} a matrix representing a coloured graph for paired data; see \code{\link{pdglasso-package}} for details.
\item \code{K}, a concentration matrix adapted to \code{pdColG}.
\item \code{sample.data}, a data frame with 60 rows and 20 columns from a multivariate normal distribution
with zero mean vector and concentration matrix \code{K}.
}
}
}
\usage{
toy_data
}
\description{
Data simulated by the function \code{\link{pdRCON.simulate}} with parameters:
\itemize{
\item \code{p=20}
\item \verb{type=c("v", "i","a)}
\item \code{force.symm=NULL}
\item \code{dens=0.3}
}
}
\keyword{datasets}
