% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy dataset generated through \code{\link{simul.pdColG}} function}
\format{
\subsection{\code{toy_data}}{

A list containing three elements:
\itemize{
\item pdColG, a \eqn{p \times p} matrix describing the coloured graphical model,
\item K, the concentration matrix associated to the model,
\item sample.data, a data frame with 60 rows and 20 columns.
}
}
}
\usage{
toy_data
}
\description{
Data simulated by the function \code{\link{simul.pdColG}} with a call:
\verb{toy_data <- pdRCON.simulate(20, type=c("v", "i","a), force.symm=NULL, dens=0.3)}
}
\keyword{datasets}
