% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy dataset generated through \code{\link[=simul.pdColG]{simul.pdColG()}} function}
\format{
\subsection{\code{toy_data}}{

A data frame with 60 rows and 20 columns.
}
}
\source{
Generated by the package functions.
}
\usage{
toy_data
}
\description{
Data simulated by using the function \link{simul.pdColG} with the following arguments:
\itemize{
\item \code{p <- 20}
\item \code{q <- p/2}
\item \verb{dens=0.3, type=c("v","i","a"), force=NULL}
}
}
\keyword{datasets}
